+++ CHANGELOG.md
@@ -3,5 +3,5 @@
 ##[0.0.1] - 2020-11-23
 ###Dev Release
 - Require Composer 1.10 or 2.0+
-- Provide two actions to install package for Symfony 4.0+ without require Symfony Recipe
+- Provide three actions to install packages (bundle, config and route) for Symfony 4.0+ without require Symfony Recipe
 - Under MIT Licence.